Once par has been worked out i then need to amend the deal so that the cards the players have reflect the par. Do this by swapping cards between players. This will be easy in some 2 player games because all cards will have been dealt if there are pre and post game cards. If there aren’t pre and post game cards or there are 3 players there will be some undealt cards
So far i have set it up so that after the deal we look at the player with the highest total hand value only,
if the difference between the hand value and the truePar is positive, meaning we need to reduce the total value of the hand, then find which player has the lowest total hand value, find it’s lowest card, now if the difference between that and the current player’s highest card value is less than the handDiff then swap it and prepare to swap again, otherwise instead of the highest card value try and select a card with value the exact right amount to reduce the handDiff to 0 if swapped with the opponent’s lowest card

This is nowhere near finished and it looks a bit too complicated and i’ve not tested what i’ve done so far either

+++ core/src/com/gmail/andrewahughes/CardButton.java
@@ -76,8 +76,8 @@ public class CardButton  extends Actor {
 
     /*when the cards are dealt to the player, all that happens is this
     * variable is set to the player's index, which will be used in the
-    * touch logic, default is 0 so we know it doesn't have a player,
-    * otherwise value will be 1, 2 or 3*/
+    * touch logic, default is -1 so we know it doesn't have a player,
+    * otherwise value will be 0,1 or 2*/
     int playerIndex =-1;
 
     /**constructor for triButton
@@ -517,7 +517,7 @@ public class CardButton  extends Actor {
      * called in the dealStage, once we have created a random list of values 0-51
      * we can assign one to this cardButton
      */
-    public void setValue(byte value){
+    public void setValue(int value){
         this.value = value;
     }
 
@@ -549,7 +549,7 @@ public class CardButton  extends Actor {
         this.text=""+this.value;
 
     }
-    public void setPlayerIndex(byte index){
+    public void setPlayerIndex(int index){
         playerIndex=index;
     }
